Fixing som Tag Interpolation Issues
