+++ src/main/java/de/neuland/jade4j/parser/Parser.java
@@ -422,7 +422,6 @@ public class Parser {
         tagNode.setLineNumber(lexer.getLineno());
         tagNode.setFileName(filename);
         tagNode.setName(name);
-        tagNode.setValue(name);
         tagNode.setBuffer(true);
         return this.tag(tagNode);
     }
